Move _update_errors fix to NetBoxModelImportForm, use NON_FIELD_ERRORS constant
